01.web.service/backend/python-services/analytics-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
kafka-python==2.0.2
python-dotenv==1.0.0
gunicorn==21.2.0
prometheus-client==0.19.0
pyjwt==2.8.0
requests==2.31.0

